breaking: Removed camelize() (unused)
