+++ packages/ember/src/index.ts
@@ -1,3 +1,2 @@
-export * from './ember/entity-name/camelize.js';
 export * from './ember/entity-name/classify.js';
 export * from './ember/entity-name/double-colonize.js';
