--- packages/ember/tests/ember/entity-name/camelize/edge-case-empty-string.test.ts
@@ -1,7 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { camelize } from '../../../../src/index.js';
-
-test('utils | ember | entity-name | camelize > edge case (empty string)', function () {
-  assert.strictEqual(camelize(''), '');
-});
