--- packages/ember/tests/ember/entity-name/camelize/edge-case-entity-name.test.ts
@@ -1,13 +1,0 @@
-import { assert, test } from '@codemod-utils/tests';
-
-import { camelize } from '../../../../src/index.js';
-
-test('utils | ember | entity-name | camelize > edge case (entity name)', function () {
-  assert.strictEqual(camelize('tracks'), 'tracks');
-  assert.strictEqual(camelize('navigation-menu'), 'navigationMenu');
-  assert.strictEqual(camelize('ui/page'), 'ui/page');
-  assert.strictEqual(
-    camelize('widgets/widget-3/tour-schedule/responsive-image'),
-    'widgets/widget3/tourSchedule/responsiveImage',
-  );
-});
